refactor: foreig keys hero
